requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.6
redis==5.0.1
django-filter==23.5
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
gunicorn==21.2.0